requirements.txt
Django==5.2
djangorestframework==3.15.2
mysqlclient==2.2.4
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.2
reportlab==4.2.2
Pillow==10.4.0
django-filter==24.2
whitenoise==6.7.0
gunicorn==22.0.0
dj-database-url==2.2.0
django-cors-headers==4.4.0
cryptography==42.0.8
